Remove seed from travis file, refactor tests and seed file
