+++ .gitignore
@@ -11,7 +11,6 @@ pids
 logs
 *.log
 public/bundle.js
-helpers/
 
 # Directory for instrumented libs generated by jscoverage/JSCover
 lib-cov
