+++ package.json
@@ -11,7 +11,7 @@
     "seed": "node database/seed.js",
     "start": "nodemon server/index.js",
     "react-dev": "webpack -d --watch",
-    "test": "mocha test/test.js"
+    "test": "mocha test/test.js --exit"
   },
   "dependencies": {
     "body-parser": "^1.18.3",
