+++ test/test.js
@@ -1,6 +1,40 @@
 const chai = require('chai');
 const chaiHttp = require('chai-http');
+const mongoose = require('mongoose');
 const app = require('../server');
+const generate = require('../helpers/generator.js');
+
+// =============== Test Database =============== \\
+
+mongoose.connect('mongodb://localhost/airbnbTest', { useNewUrlParser: true });
+
+const testSchema = new mongoose.Schema({
+  title: String,
+  city: String,
+  numGuests: Number,
+  numBedrooms: Number,
+  numBeds: Number,
+  numBaths: Number,
+  description: String
+});
+
+const TestDesc = mongoose.model('TestDesc', testSchema);
+
+const noun = generate.genNoun();
+const numGuests = generate.genNumGuests();
+const numBedrooms = generate.genNumBedrooms();
+
+const testHome = new TestDesc({
+  title: generate.genTitle(noun),
+  city: generate.genCity(),
+  numGuests,
+  numBedrooms,
+  numBeds: generate.genNumBeds(numGuests),
+  numBaths: generate.genNumBaths(),
+  description: generate.genDescription(noun, numBedrooms)
+});
+
+// =============== End Database =============== \\
 
 chai.use(chaiHttp);
 chai.should();
@@ -37,111 +71,73 @@ describe('Server', () => {
 });
 
 describe('Database', () => {
-  it('should return status 200 from GET /description', (done) => {
-    chai.request(app)
-      .get('/description')
-      .end((err, res) => {
-        res.should.have.status(200);
-        done();
-      });
+  before(() => {
+    testHome.save((err) => {
+      if (err) {
+        console.log('Error saving home description to test database:', err);
+      } else {
+        console.log('Saved test home:', testHome.title);
+      }
+    });
   });
 
-  it('should return a description object', (done) => {
-    chai.request(app)
-      .get('/description')
-      .query({ _id: randNum() })
-      .end((err, res) => {
+  it('should return a description object', () => {
+    TestDesc.find({}, (err, res) => {
+      if (err) console.log('Error reading from db:', err);
+      else {
         res.body[0].should.be.an('object');
-        done();
-      });
+      }
+    });
   });
 
   describe('Description Object', () => {
-    it('should have a numeric id', (done) => {
-      chai.request(app)
-        .get('/description')
-        .query({ _id: randNum() })
-        .end((err, res) => {
-          res.body[0]._id.should.be.a('number');
-          done();
-        });
+    it('should have a title', () => {
+      TestDesc.find({}, (err, res) => {
+        res.body[0].title.should.exist;
+        res.body[0].title.should.be.a('string');
+      });
     });
 
-    it('should have a title', (done) => {
-      chai.request(app)
-        .get('/description')
-        .query({ _id: randNum() })
-        .end((err, res) => {
-          res.body[0].title.should.exist;
-          res.body[0].title.should.be.a('string');
-          done();
-        });
+    it('should have a city', () => {
+      TestDesc.find({}, (err, res) => {
+        res.body[0].city.should.exist;
+        res.body[0].city.should.be.a('string');
+      });
     });
 
-    it('should have a city', (done) => {
-      chai.request(app)
-        .get('/description')
-        .query({ _id: randNum() })
-        .end((err, res) => {
-          res.body[0].city.should.exist;
-          res.body[0].city.should.be.a('string');
-          done();
-        });
+    it('should have a number of allowed guests', () => {
+      TestDesc.find({}, (err, res) => {
+        res.body[0].numGuests.should.exist;
+        res.body[0].numGuests.should.be.a('number');
+      });
     });
 
-    it('should have a number of allowed guests', (done) => {
-      chai.request(app)
-        .get('/description')
-        .query({ _id: randNum() })
-        .end((err, res) => {
-          res.body[0].numGuests.should.exist;
-          res.body[0].numGuests.should.be.a('number');
-          done();
-        });
+    it('should have a number of bedrooms', () => {
+      TestDesc.find({}, (err, res) => {
+        res.body[0].numBedrooms.should.exist;
+        res.body[0].numBedrooms.should.be.a('number');
+      });
     });
 
-    it('should have a number of bedrooms', (done) => {
-      chai.request(app)
-        .get('/description')
-        .query({ _id: randNum() })
-        .end((err, res) => {
-          res.body[0].numBedrooms.should.exist;
-          res.body[0].numBedrooms.should.be.a('number');
-          done();
-        });
+    it('should have a number of beds', () => {
+      TestDesc.find({}, (err, res) => {
+        res.body[0].numBeds.should.exist;
+        res.body[0].numBeds.should.be.a('number');
+      });
     });
 
-    it('should have a number of beds', (done) => {
-      chai.request(app)
-        .get('/description')
-        .query({ _id: randNum() })
-        .end((err, res) => {
-          res.body[0].numBeds.should.exist;
-          res.body[0].numBeds.should.be.a('number');
-          done();
-        });
+    it('should have a number of bathrooms', () => {
+      TestDesc.find({}, (err, res) => {
+        res.body[0].numBaths.should.exist;
+        res.body[0].numBaths.should.be.a('number');
+      });
     });
 
-    it('should have a number of bathrooms', (done) => {
-      chai.request(app)
-        .get('/description')
-        .query({ _id: randNum() })
-        .end((err, res) => {
-          res.body[0].numBaths.should.exist;
-          res.body[0].numBaths.should.be.a('number');
-          done();
-        });
-    });
-
-    it('should have a description paragraph', (done) => {
-      chai.request(app)
-        .get('/description')
-        .query({ _id: randNum() })
-        .end((err, res) => {
-          res.body[0].description.should.exist;
-          res.body[0].description.should.be.a('string');
-          done();
-        });
+    it('should have a description paragraph', () => {
+      TestDesc.find({}, (err, res) => {
+        res.body[0].description.should.exist;
+        res.body[0].description.should.be.a('string');
+      });
     });
   });
 });
